Create compare_solutions module
